Port Animation to YAML based enum

+++ src/map/packets/s2c/0x030_effect.cpp
@@ -31,5 +31,5 @@ GP_SERV_COMMAND_EFFECT::GP_SERV_COMMAND_EFFECT(const CCharEntity* PChar, const S
     packet.ActIndex  = PChar->targid;
     packet.EffectNum = effect;
     packet.Type      = param;
-    packet.Status    = PChar->animation;
+    packet.Status    = static_cast<int8_t>(PChar->animation);
 }
